+++ backend/libs/room-operations-crud.ts
@@ -6,10 +6,8 @@ import {
     Watcher,
     WatcherState,
 } from '../../extension/src/types';
-import { dynamoDB } from '../libs/dynamodb-utils';
+import { dynamoDB } from './dynamodb-utils';
 import { marshallRoom, unmarshallRoom } from './room-marshalling';
-import { assignRoomIntent } from './room-utils';
-import { updateWatcher } from './watcher-operations';
 
 /**
  * Find a room by ID in DDB
@@ -138,57 +136,5 @@ export const updateRoom = async (
     } catch (e) {
         console.error('Failed update room', e);
         throw new Error('Failed to update room');
-    }
-};
-
-/**
- * Add a new user to the room's "watchers" map
- */
-export const joinExistingRoom = async (
-    room: Room,
-    tableName: string,
-    watcherConnectionString: string,
-    dynamoDb: DocumentClient,
-): Promise<Room> => {
-    const newWatcher: Watcher = {
-        id: watcherConnectionString,
-        connectionId: watcherConnectionString,
-        joinedAt: new Date(),
-        lastVideoTimestamp: null,
-        lastHeartbeat: new Date(),
-        currentVideoStatus: WatcherState.UNKNOWN,
-        initialSync: false,
-        userAgent: 'TODO',
-    };
-    return updateWatcher(room, tableName, newWatcher, dynamoDb);
-};
-
-/**
- * Checks the user permission to modify the room the syncIntent
- * Update it
- *
- * @param room The room to update
- * @param tableName The name of the dynamodb table
- * @param watcherConnectionString The connection id of the user trying to update the state
- * @param syncIntent The new syncIntent for the room
- */
-export const updateRoomSyncIntent = async (
-    room: Room,
-    tableName: string,
-    syncIntent: SyncIntent,
-): Promise<Room> => {
-    room = assignRoomIntent(room, syncIntent);
-    console.log('Update the room with the new sync intent');
-    return updateRoom(room, tableName);
-};
-
-export const ensureOnlyOwnerCanDoThisError = (
-    room: Room,
-    watcher: string,
-): void => {
-    if (watcher != room.ownerId) {
-        throw new Error(
-            `Watcher ${watcher} is not authorized to perform this operation, only ${room.ownerId} can`,
-        );
     }
 };
